inline useDpi hook
